Better error handling telemetry, select aggregation
The error handling for plenary.curl did not work, because it is an async
call. plenary.curl documentation is not easy, but I found this
https://github.com/nvim-lua/plenary.nvim/pull/471
and included on_error to handle possible errors

When the user forgets the aggregation key, a selection is called.
As it is vim.ui.select, it should be with the picker of choice